[Refactor][Kernel] Flatten kernels/reduction/ subdirectories
Collapse each single-fwd-file subdirectory into a flat module. Split
softmax/ into separate softmax.py and logsumexp.py. Update all imports
and manifest paths.

Part 5 of #924.
