+++ tileops/kernels/reduction/__init__.py
@@ -17,8 +17,9 @@ from ._primitives import (
 from .argreduce import ArgreduceKernel
 from .cumulative import CumulativeKernel
 from .logical_reduce import LogicalReduceKernel
+from .logsumexp import LogSumExpKernel
 from .reduce import ReduceKernel
-from .softmax import LogSumExpKernel, SoftmaxKernel
+from .softmax import SoftmaxKernel
 
 # Placeholder imports for reduction kernels.
 # Each sub-category PR uncomments its own lines.
