--- tileops/kernels/reduction/argreduce/__init__.py
@@ -1,7 +1,0 @@
-"""Argreduce kernel subpackage (argmax, argmin)."""
-
-from .fwd import ArgreduceKernel
-
-__all__: list[str] = [
-    "ArgreduceKernel",
-]
