--- tileops/kernels/reduction/cumulative/__init__.py
@@ -1,7 +1,0 @@
-"""Cumulative scan kernel subpackage (cumsum, cumprod)."""
-
-from .fwd import CumulativeKernel
-
-__all__: list[str] = [
-    "CumulativeKernel",
-]
