--- tileops/kernels/reduction/logical_reduce/__init__.py
@@ -1,7 +1,0 @@
-"""Logical reduce kernel subpackage (any, all, count_nonzero)."""
-
-from .fwd import LogicalReduceKernel
-
-__all__: list[str] = [
-    "LogicalReduceKernel",
-]
