--- tileops/kernels/reduction/reduce/__init__.py
@@ -1,7 +1,0 @@
-"""Reduce kernel subpackage (sum, mean, amin, amax, prod, std, var, var_mean)."""
-
-from .fwd import ReduceKernel
-
-__all__: list[str] = [
-    "ReduceKernel",
-]
