--- tileops/kernels/reduction/softmax/__init__.py
@@ -1,10 +1,0 @@
-# Copyright (c) Tile-AI. All rights reserved.
-"""Softmax-family kernels (softmax, log_softmax, logsumexp)."""
-
-from .logsumexp_fwd import LogSumExpKernel
-from .softmax_fwd import SoftmaxKernel
-
-__all__: list[str] = [
-    "LogSumExpKernel",
-    "SoftmaxKernel",
-]
