--- tileops/kernels/reduction/vector_norm/__init__.py
@@ -1,7 +1,0 @@
-"""Vector norm kernel subpackage (l1, l2, inf norm)."""
-
-from .fwd import VectorNormKernel
-
-__all__: list[str] = [
-    "VectorNormKernel",
-]
