+++ tileops/ops/reduction/logsumexp.py
@@ -14,7 +14,7 @@ from typing import Dict, List, Optional, Union
 import torch
 
 from tileops.kernels.kernel import Kernel
-from tileops.kernels.reduction.softmax import LogSumExpKernel
+from tileops.kernels.reduction.logsumexp import LogSumExpKernel
 
 from ._softmax_base import _SoftmaxBaseOp
